refactor conversation: SimpleAIResponseGenerator 분리
