+++ src/main/java/com/anyang/maruni/domain/conversation/infrastructure/ai/OpenAIResponseAdapter.java
@@ -1,7 +1,4 @@
-package com.anyang.maruni.domain.conversation.infrastructure;
-
-import java.util.List;
-import java.util.Map;
+package com.anyang.maruni.domain.conversation.infrastructure.ai;
 
 import org.springframework.ai.chat.model.ChatModel;
 import org.springframework.ai.chat.model.ChatResponse;
@@ -11,24 +8,25 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
-import com.anyang.maruni.domain.conversation.domain.entity.EmotionType;
+import com.anyang.maruni.domain.conversation.domain.port.AIResponsePort;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 /**
- * 간단한 AI 응답 생성기 (MVP 버전 - Spring AI 구현)
+ * OpenAI 기반 AI 응답 생성기
  *
- * Spring AI를 사용하여 사용자 메시지에 대한 AI 응답을 생성합니다.
- * MVP에서는 기본적인 프롬프트와 간단한 감정 분석만 제공합니다.
+ * 기존 SimpleAIResponseGenerator의 AI 응답 생성 로직을 OpenAI 전용으로 분리했습니다.
+ * Port-Adapter 패턴을 통해 AI 모델 변경에 대비한 독립적인 구조를 제공합니다.
  */
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class SimpleAIResponseGenerator {
+public class OpenAIResponseAdapter implements AIResponsePort {
 
     private final ChatModel chatModel;
 
+    // 기존 SimpleAIResponseGenerator의 설정값 그대로 사용
     @Value("${spring.ai.openai.chat.options.model}")
     private String model;
 
@@ -38,21 +36,15 @@ public class SimpleAIResponseGenerator {
     @Value("${spring.ai.openai.chat.options.max-tokens}")
     private Integer maxTokens;
 
-    // 응답 관련 상수
+    // 응답 관련 상수 (기존 값 그대로 유지)
     private static final int MAX_RESPONSE_LENGTH = 100;
     private static final int ELLIPSIS_LENGTH = 3;
     private static final String ELLIPSIS = "...";
 
-    // 메시지 관련 상수
+    // 메시지 관련 상수 (기존 값 그대로 유지)
     private static final String DEFAULT_USER_MESSAGE = "안녕하세요";
     private static final String DEFAULT_RESPONSE = "안녕하세요! 어떻게 지내세요?";
     private static final String SYSTEM_PROMPT = "당신은 노인 돌봄 전문 AI 상담사입니다. 따뜻하고 공감적으로 30자 이내로 응답하세요.";
-
-    // 감정분석 키워드 맵
-    private static final Map<EmotionType, List<String>> EMOTION_KEYWORDS = Map.of(
-        EmotionType.NEGATIVE, List.of("슬프", "우울", "아프", "힘들", "외로", "무서", "걱정", "답답"),
-        EmotionType.POSITIVE, List.of("좋", "행복", "기쁘", "감사", "즐거", "만족", "고마")
-    );
 
     /**
      * 사용자 메시지에 대한 AI 응답 생성
@@ -60,14 +52,15 @@ public class SimpleAIResponseGenerator {
      * @param userMessage 사용자 메시지
      * @return AI 응답 내용
      */
+    @Override
     public String generateResponse(String userMessage) {
         try {
             log.info("AI 응답 생성 요청: {}", userMessage);
 
-            // 입력 검증
+            // 입력 검증 (기존 로직 그대로)
             String sanitizedMessage = sanitizeUserMessage(userMessage);
 
-            // Spring AI로 응답 생성
+            // Spring AI로 응답 생성 (기존 로직 그대로)
             String response = callSpringAI(sanitizedMessage);
             String finalResponse = truncateResponse(response);
 
@@ -80,40 +73,7 @@ public class SimpleAIResponseGenerator {
     }
 
     /**
-     * 간단한 감정 분석 수행 (MVP: 키워드 기반)
-     *
-     * @param message 분석할 메시지
-     * @return 감정 타입
-     */
-    public EmotionType analyzeBasicEmotion(String message) {
-        log.debug("감정 분석 시작: {}", message);
-
-        // null 또는 빈 문자열 처리
-        if (!StringUtils.hasText(message)) {
-            return EmotionType.NEUTRAL;
-        }
-
-        String lowerMessage = message.toLowerCase();
-
-        // 부정적 키워드 체크 (우선 순위 높음)
-        if (containsAnyKeyword(lowerMessage, EMOTION_KEYWORDS.get(EmotionType.NEGATIVE))) {
-            log.debug("부정적 감정 감지: NEGATIVE");
-            return EmotionType.NEGATIVE;
-        }
-
-        // 긍정적 키워드 체크
-        if (containsAnyKeyword(lowerMessage, EMOTION_KEYWORDS.get(EmotionType.POSITIVE))) {
-            log.debug("긍정적 감정 감지: POSITIVE");
-            return EmotionType.POSITIVE;
-        }
-
-        // 기본값: 중립
-        log.debug("중립적 감정: NEUTRAL");
-        return EmotionType.NEUTRAL;
-    }
-
-    /**
-     * 사용자 메시지 입력 검증 및 정제
+     * 사용자 메시지 입력 검증 및 정제 (기존 로직 그대로)
      */
     private String sanitizeUserMessage(String userMessage) {
         if (!StringUtils.hasText(userMessage)) {
@@ -123,7 +83,7 @@ public class SimpleAIResponseGenerator {
     }
 
     /**
-     * Spring AI를 사용한 응답 생성
+     * Spring AI를 사용한 응답 생성 (기존 로직 그대로)
      */
     private String callSpringAI(String userMessage) {
         // 시스템 프롬프트와 사용자 메시지를 결합한 프롬프트 생성
@@ -144,7 +104,7 @@ public class SimpleAIResponseGenerator {
     }
 
     /**
-     * 응답 길이 제한 (SMS 특성상)
+     * 응답 길이 제한 (SMS 특성상) - 기존 로직 그대로
      */
     private String truncateResponse(String response) {
         if (response.length() > MAX_RESPONSE_LENGTH) {
@@ -154,18 +114,10 @@ public class SimpleAIResponseGenerator {
     }
 
     /**
-     * API 에러 처리
+     * API 에러 처리 (기존 로직 그대로)
      */
     private String handleApiError(Exception e) {
         log.error("AI 응답 생성 실패: {}", e.getMessage(), e);
         return DEFAULT_RESPONSE;
-    }
-
-    /**
-     * 메시지에 키워드 목록 중 하나라도 포함되는지 확인
-     */
-    private boolean containsAnyKeyword(String message, List<String> keywords) {
-        return keywords.stream()
-                .anyMatch(message::contains);
     }
 }
